Got sockets partially working
Rename Socket class, fix bug involving Door model parceling
